+++ packages/lwc-engine/jest.config.js
@@ -8,4 +8,5 @@ module.exports = {
     moduleNameMapper: {
         'test-utils': path.resolve(__dirname, 'scripts/jest/test-utils.js'),
     },
+    setupTestFrameworkScriptFile: path.resolve(__dirname, 'scripts/jest/setup-test.js'),
 };