+++ packages/lwc-engine/src/framework/decorators/__tests__/api.spec.ts
@@ -109,7 +109,6 @@ describe('decorators/api.ts', () => {
             expect(elm.x).toBe('foo');
             expect(elm.textContent).toBe('foo');
         });
-
     });
 
     describe('@api get/set x', () => {