+++ packages/lwc-engine/src/framework/decorators/__tests__/track.spec.ts
@@ -192,7 +192,6 @@ describe('track.ts', () => {
                 document.body.appendChild(elm);
             }).toThrow();
         });
-
     });
 
     describe('@track regression', () => {